invocacion  notifySearchBiography() + rename+
Modificamos nombre del paquete, creamos una nueva funcion notifySearchBiography
