+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".moredetails.view.OtherDetailsViewActivity"
+        <activity android:name=".otherdetails.view.OtherDetailsViewActivity"
            android:theme="@style/Theme.AppCompat.NoActionBar"/>
     </application>
