--- app/src/main/java/ayds/lisboa/songinfo/moredetails/controller/OtherDetailsControllerInjector.kt
@@ -1,12 +1,0 @@
-package ayds.lisboa.songinfo.moredetails.controller
-
-import ayds.lisboa.songinfo.moredetails.model.OtherDetailsModelInjector
-import ayds.lisboa.songinfo.moredetails.view.OtherDetailsView
-
-object OtherDetailsControllerInjector {
-    fun onViewStarted(otherDetailsView: OtherDetailsView) {
-        OtherDetailsControllerImpl(OtherDetailsModelInjector.getOtherDetailsModel()).apply {
-            setOtherDetailsView(otherDetailsView)
-        }
-    }
-}
