--- app/src/main/java/ayds/lisboa/songinfo/moredetails/model/entities/LastFMArtistBiography.kt
@@ -1,25 +1,0 @@
-package ayds.lisboa.songinfo.moredetails.model.entities
-
-interface ArtistBiography{
-    val id : String
-    val artist: String
-    val biography: String
-    val url: String
-    var isLocallyStored: Boolean
-}
-
-data class LastFMArtistBiography(
-    override val id: String,
-    override val artist: String,
-    override val biography: String,
-    override val url: String,
-    override var isLocallyStored: Boolean = false
-): ArtistBiography {}
-
-object EmptyArtistBiography : ArtistBiography {
-    override val id: String = ""
-    override val artist: String = ""
-    override val biography: String = ""
-    override val url: String = ""
-    override var isLocallyStored: Boolean = false
-}
