--- app/src/main/java/ayds/lisboa/songinfo/moredetails/model/repository/external/lastfm/LastFMInjector.kt
@@ -1,6 +1,0 @@
-package ayds.lisboa.songinfo.moredetails.model.repository.external.lastfm
-
-object LastFMInjector {
-
-    val lastFMService : LastFMService = LastFMArtistInjector.lastFMArtistService
-}
