--- app/src/main/java/ayds/lisboa/songinfo/moredetails/view/OtherDetailsUiEvent.kt
@@ -1,6 +1,0 @@
-package ayds.lisboa.songinfo.moredetails.view
-
-sealed class OtherDetailsUiEvent {
-    object SearchBiography: OtherDetailsUiEvent() //TODO cual es la funcionalidad?
-    object OpenBiographyArticleUrl : OtherDetailsUiEvent()
-}
