+++ app/src/main/java/ayds/lisboa/songinfo/otherdetails/controller/OtherDetailsControllerInjector.kt
@@ -1,0 +1,12 @@
+package ayds.lisboa.songinfo.otherdetails.controller
+
+import ayds.lisboa.songinfo.otherdetails.model.OtherDetailsModelInjector
+import ayds.lisboa.songinfo.otherdetails.view.OtherDetailsView
+
+object OtherDetailsControllerInjector {
+    fun onViewStarted(otherDetailsView: OtherDetailsView) {
+        OtherDetailsControllerImpl(OtherDetailsModelInjector.getOtherDetailsModel()).apply {
+            setOtherDetailsView(otherDetailsView)
+        }
+    }
+}
