+++ app/src/main/java/ayds/lisboa/songinfo/otherdetails/model/repository/external/lastfm/LastFMAPI.java
@@ -1,0 +1,13 @@
+package ayds.lisboa.songinfo.otherdetails.model.repository.external.lastfm;
+
+
+import retrofit2.Call;
+import retrofit2.http.GET;
+import retrofit2.http.Query;
+
+public interface LastFMAPI {
+
+  @GET("?method=artist.getinfo&api_key=0a657854db69e551c97d541ca9ebcef4&format=json")
+  Call<String> getArtistInfo(@Query("artist") String artist);
+
+}
