+++ app/src/main/java/ayds/lisboa/songinfo/otherdetails/model/repository/local/lastfm/LastFMLocalStorage.kt
@@ -1,0 +1,10 @@
+package ayds.lisboa.songinfo.otherdetails.model.repository.local.lastfm
+
+import ayds.lisboa.songinfo.otherdetails.model.entities.LastFMArtistBiography
+
+interface LastFMLocalStorage {
+
+    fun saveArtist(artistBiography: LastFMArtistBiography)
+
+    fun getInfo(artist: String): LastFMArtistBiography?
+}
