+++ app/src/main/java/ayds/lisboa/songinfo/otherdetails/view/OtherDetailsUiEvent.kt
@@ -1,0 +1,6 @@
+package ayds.lisboa.songinfo.otherdetails.view
+
+sealed class OtherDetailsUiEvent {
+    object SearchBiography: OtherDetailsUiEvent() //TODO cual es la funcionalidad?
+    object OpenBiographyArticleUrl : OtherDetailsUiEvent()
+}
