+++ app/src/main/java/ayds/lisboa/songinfo/otherdetails/view/OtherDetailsUiState.kt
@@ -1,0 +1,13 @@
+package ayds.lisboa.songinfo.otherdetails.view
+
+data class OtherDetailsUiState (
+    val artistName: String = "",
+    val viewFullArticleUrl: String = "",
+    val artistBiographyText: String = "",
+    val actionsEnabled: Boolean = false,
+) {
+
+    companion object {
+        const val IMAGE_URL_SERVICE = "https://upload.wikimedia.org/wikipedia/commons/thumb/d/d4/Lastfm_logo.svg/320px-Lastfm_logo.svg.png"
+    }
+}
