+++ app/src/main/java/ayds/lisboa/songinfo/otherdetails/view/OtherDetailsViewInjector.kt
@@ -1,0 +1,14 @@
+package ayds.lisboa.songinfo.otherdetails.view
+
+import ayds.lisboa.songinfo.otherdetails.controller.OtherDetailsControllerInjector
+import ayds.lisboa.songinfo.otherdetails.model.OtherDetailsModelInjector
+
+object OtherDetailsViewInjector {
+
+    val biographyDescriptionHelper: BiographyDescriptionHelper = BiographyDescriptionHelperImpl()
+
+    fun init(otherDetailsView: OtherDetailsView) {
+        OtherDetailsModelInjector.initOtherDetailsModel(otherDetailsView)
+        OtherDetailsControllerInjector.onViewStarted(otherDetailsView)
+    }
+}
